refactor: move RejectsUnknownKeys into a shared write base module
The pipeline façade's bodies need the same rule, and importing it from the
chatbot settings serializers would be the wrong dependency.
