+++ apps/api/v2/write/serializers.py
@@ -16,35 +16,13 @@ from django.db import transaction
 from drf_spectacular.utils import extend_schema_serializer
 from rest_framework import serializers
 
+from apps.api.v2.write.base import RejectsUnknownKeys
 from apps.api.v2.write.fields import NfcCharField, OptionalTextField, TeamScopedRelatedField
 from apps.experiments.helpers import excluded_voice_services
 from apps.experiments.models import ConsentForm, Experiment, SyntheticVoice
 from apps.pipelines.models import Pipeline
 from apps.service_providers.models import TraceProvider, VoiceProvider
 from apps.service_providers.utils import get_first_llm_provider_by_team, get_first_llm_provider_model
-
-
-class RejectsUnknownKeys:
-    """Refuse a request body carrying keys this serializer does not declare.
-
-    DRF's default is to drop them silently, which for a human is a typo they spot in the echoed
-    response and for an agent is a 200 that wrote nothing. The consumer here is an agent, so a
-    misspelled key has to be an error it can act on.
-
-    Hooked into ``to_internal_value`` rather than ``validate`` because that is where a serializer
-    sees its own raw input wherever it is mounted. A ``validate``-based check would have to read
-    ``initial_data``, which DRF sets on the root serializer alone.
-    """
-
-    def to_internal_value(self, data: Any) -> Any:
-        if isinstance(data, dict):
-            unknown = sorted(set(data) - set(self.fields))
-            if unknown:
-                accepted = ", ".join(sorted(self.fields))
-                raise serializers.ValidationError(
-                    {key: f"Unrecognised field. Accepted here: {accepted}." for key in unknown}
-                )
-        return super().to_internal_value(data)
 
 
 class ChatbotCreateSerializer(RejectsUnknownKeys, serializers.Serializer):
